app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
httpx
openai
chromadb
pypdf
python-docx
python-dotenv
